separate delete functionality to different files
